AI Quotes generator added and Improved Folder Stucture
