+++ Ai Quotes Generator/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>AI Quotes Generator</title>
+    <link rel="stylesheet" href="styles.css">
+</head>
+<body>
+    <header>
+        <h1>AI Quotes Generator</h1>
+    </header>
+    <div class="container">
+        <div id="quoteContainer">
+            <p id="quoteText"></p>
+        </div>
+        <button id="getQuoteBtn" style="background-color: 3498db;">Get Quote</button>
+    </div>
+    <script src="script.js"></script>
+</body>
+</html>
